requirements.txt
Flask
MarkupSafe
multidict
numpy
openai==0.28.0
opencv-python
openpyxl
pandas
pillow
pillow_heif
pypinyin
python-dotenv
webdriver_manager==4.0.1
selenium==4.23.0
requests
Werkzeug
msal
gunicorn
opencv-python-headless

